Moving Detection transforms to references.
